requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
apscheduler==3.10.4
anthropic==0.49.0
httpx==0.28.1
aiosqlite==0.20.0
python-dotenv==1.0.1
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.10.6
